Add usage of atoms and molucules to separate common components
- Fix responsiveness of the Profile component
- Remove horizontal gutter for larger viewports
