--- src/components/common/Profile/Profile.module.scss
@@ -1,20 +1,0 @@
-.profile {
-    :global {
-        .picture {
-            background-image: url('../../../assets/img/profile.jpg');
-            background-size: cover;
-            border-radius: 50%;
-            -webkit-border-radius: 50%;
-            -moz-border-radius: 50%;
-            display: block;
-            width: 100%;
-            height: 0;
-            padding-bottom: 100%;
-        }
-
-        .text {
-            @include responsive('font-size', '14px', '16px', '18px', '20px', '25px');
-            @include responsive('margin-top', '25px', '30px', '55px', '55px', '80px');
-        }
-    } 
-}
