+++ src/components/common/atoms/Snow/Snow.module.scss
