+++ src/components/common/molecules/Profile/Profile.module.scss
@@ -1,0 +1,25 @@
+.profile {
+    :global {
+        .picture-row {
+            @include responsive('width', '170px', '170px', '200px', '200px', '200px');
+
+            .picture {
+                background-image: url('../../../../assets/img/profile.jpg');
+                background-size: cover;
+                border-radius: 50%;
+                -webkit-border-radius: 50%;
+                -moz-border-radius: 50%;
+                display: block;
+                width: 100%;
+                height: 0;
+                padding-bottom: 100%;
+            }
+        }
+
+        .text {
+            font-size: 16px;
+            max-width: 650px;
+            @include responsive('margin-top', '25px', '30px', '55px', '55px', '80px');
+        }
+    } 
+}
