+++ src/index.scss
@@ -37,6 +37,7 @@ section {
 
 .title {
   font-family: Raleway;
-  font-weight: 600;
+  font-weight: 300;
   text-align: center;
+  letter-spacing: 1px;
 }
